+++ elfi/__init__.py
@@ -14,6 +14,7 @@ from elfi.client import get_client, set_client
 from elfi.methods.diagnostics import TwoStageSelection
 from elfi.methods.model_selection import *
 from elfi.methods.inference.bolfi import *
+from elfi.methods.inference.bolfire import *
 from elfi.methods.inference.romc import *
 from elfi.methods.inference.samplers import *
 from elfi.methods.post_processing import adjust_posterior
@@ -30,4 +31,4 @@ __author__ = 'ELFI authors'
 __email__ = 'elfi-support@hiit.fi'
 
 # make sure __version_ is on the last non-empty line (read by setup.py)
-__version__ = '0.8.2'
+__version__ = '0.8.3'